+++ baranov.pavel/P4/inputMatrix.hpp
@@ -1,0 +1,8 @@
+#ifndef INPUT_MATRIX_HPP
+#define INPUT_MATRIX_HPP
+#include <iostream>
+
+void inputMatrix(std::istream & input, int * matrix, size_t size);
+
+#endif
+
