+++ baranov.pavel/P4/maxSumDiagonal.hpp
@@ -1,0 +1,6 @@
+#ifndef MAX_SUM_DIAGONAL_HPP
+#define MAX_SUM_DIAGONAL_HPP
+
+int maxSumDiagonal(int * matrix, unsigned int rows, unsigned int columns);
+
+#endif
